# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/rooms/{slug}": {
      "get": {
        "operationId": "RoomsController_getRoomBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/create-brand": {
      "post": {
        "operationId": "RoomsController_createBrand",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{roomId}/connect-brand/{brandId}": {
      "patch": {
        "operationId": "RoomsController_connectBrandToRoom",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/create": {
      "post": {
        "operationId": "RoomsController_createRoom",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Rooms"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bluetree API",
    "description": "API for managing Bluetree resources",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateTeamMemberDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "profileImageUrl": {
            "type": "string"
          }
        },
        "required": [
          "fullName",
          "role",
          "profileImageUrl"
        ]
      },
      "CreateContactInfoDto": {
        "type": "object",
        "properties": {
          "linkedin": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          },
          "telegram": {
            "type": "string"
          }
        }
      },
      "CreateAboutUsDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        },
        "required": [
          "content"
        ]
      },
      "CreateBrandDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "teamMembers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTeamMemberDto"
            }
          },
          "contactInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateContactInfoDto"
            }
          },
          "aboutUs": {
            "$ref": "#/components/schemas/CreateAboutUsDto"
          }
        },
        "required": [
          "name",
          "teamMembers",
          "contactInfos",
          "aboutUs"
        ]
      },
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "brandId": {
            "type": "number"
          }
        },
        "required": [
          "slug"
        ]
      }
    }
  }
}